CA-151389: Fix retVal issue for functional HBA tests.
The 'retVal' variable gets set to 'False' when the last scsi host
scanned is empty. This variable is also used at other places and
returned from this function to determine the 'PASS/FAIL' status of the
tests.

This patch resolves this problem by using a different
variable to capture the state of a LUN.

Signed-off-by: Siddharth Vinothkumar <siddharth.vinothkumar@citrix.com>
Reviewed-by: Germano Percossi <germano.percossi@citrix.com>

GitHub: closes #35 on xenserver/xencert